minor correction
